app/requirements.txt
psycopg2==2.9.5
djangorestframework==3.11.2
django-filter==2.4.0
channels==3.0.1
channels-redis==3.4.1
daphne==3.0.2
pytz==2022.7.1
Django==3.1.14
Markdown==3.2.1
scipy==1.10.1
backoff==2.2.1
markdown==3.2.1
PyYAML==6.0
aioredis==1.3.1
bleach==6.0.0
redis==4.5.4
python-box==7.0.1
regex==2023.5.5
python-dateutil==2.8.2
watchdog==3.0.0
boto3>=1.28.57
awscli>=1.29.57
botocore>=1.31.57
celery[redis]>=5.0,<6
ipdb
openai>=0.27.1
langchain
langchain-openai>=0.1.7



